# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PharmSci E-Voting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerVoter
      tags: [auth]
      summary: Register a student voter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoterRegistrationInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: loginVoter
      tags: [auth]
      summary: Login on voting day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoterLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoterSession"
        "401":
          description: Invalid credentials or locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getVoterSession
      tags: [auth]
      summary: Get current voter session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoterSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Ballot ──────────────────────────────────────────────────────────────────
...
  /admin/voters:
    get:
      operationId: listVoters
      tags: [admin-voters]
      summary: List all registered voters
      parameters:
        - name: level
          in: query
          schema:
            type: string
        - name: voted
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Voters list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoterListResponse"
  /admin/voters/{id}:
    patch:
      operationId: updateVoter
      tags: [admin-voters]
      summary: Update voter record (name or level only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoterUpdate"
      responses:
        "200":
          description: Updated voter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voter"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteVoter
      tags: [admin-voters]
      summary: Invalidate/delete a voter registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/student-records:
    get:
      operationId: listStudentRecords
      tags: [admin-students]
      summary: List official student records used for validation
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Student records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecordListResponse"
    post:
      operationId: importStudentRecords
      tags: [admin-students]
      summary: Bulk import student records (JSON array)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentRecordImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ApiError:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required: [error]

    # ── Election ────────────────────────────────────────────────────────────────
...
    VoterRegistrationInput:
      type: object
      properties:
        matricNumber:
          type: string
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        level:
          type: string
          enum: ["100L", "200L", "300L", "400L", "500L", "600L"]
      required: [matricNumber, email, password, fullName, level]
    RegistrationResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        matricNumber:
          type: string
        fullName:
          type: string
        level:
          type: string
      required: [success, message, matricNumber, fullName, level]
    VoterLoginInput:
      type: object
      properties:
        matricNumber:
          type: string
        password:
          type: string
      required: [matricNumber, password]
    VoterSession:
      type: object
      properties:
        id:
          type: integer
        matricNumber:
          type: string
        fullName:
          type: string
        level:
          type: string
        hasVoted:
          type: boolean
      required: [id, matricNumber, fullName, level, hasVoted]

    # ── Ballot ──────────────────────────────────────────────────────────────────
...
    Voter:
      type: object
      properties:
        id:
          type: integer
        matricNumber:
          type: string
        fullName:
          type: string
        email:
          type: string
        level:
          type: string
        hasVoted:
          type: boolean
        registrationTimestamp:
          type: string
        ipAddress:
          type: ["string", "null"]
        voteTimestamp:
          type: ["string", "null"]
      required: [id, matricNumber, fullName, email, level, hasVoted, registrationTimestamp]
    VoterUpdate:
      type: object
      properties:
        fullName:
          type: string
        level:
          type: string
          enum: ["100L", "200L", "300L", "400L", "500L", "600L"]
    VoterListResponse:
      type: object
      properties:
        voters:
          type: array
          items:
            $ref: "#/components/schemas/Voter"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [voters, total, page, limit]

    # ── Offices & Candidates ─────────────────────────────────────────────────────
...
    StudentRecord:
      type: object
      properties:
        id:
          type: integer
        matricNumber:
          type: string
        email:
          type: string
        fullName:
          type: string
        level:
          type: string
        isRegistered:
          type: boolean
      required: [id, matricNumber, email, fullName, level, isRegistered]
    StudentRecordListResponse:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/StudentRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [students, total, page, limit]
    StudentRecordImport:
      type: object
      properties:
        records:
          type: array
          items:
            type: object
            properties:
              matricNumber:
                type: string
              email:
                type: string
              fullName:
                type: string
              level:
                type: string
            required: [matricNumber, email, fullName, level]
      required: [records]
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: integer
        message:
          type: string
      required: [imported, skipped, errors, message]